* origin/v38r3-patches: (62 commits)
  Bump version to 38.3 and update change log
  listcomponents: skip factories from other libraries
  Fix building with Catch2 v3
  GaudiTestSuite: pre-commit fixes
  Add mutex to protect accesses to PersistencySvc::m_cnvServices
  Add test to expose race condition in PersistencySvc
  Minor clean up of PersistencySvc
  Minor improvement to PersistencySvc messages
  Tweak BranchWrapper implementation
  Add hard limit on test stdout/err size
  Restore behavior of GAUDI-976 test
  use string_view consistently in parse methods
  Fix a number of type-comparison (E721) violations
  Small reformatting of Python code (black->ruff)
  Change in Python import formatting (isort->ruff)
  Replace flake8, black and isort with ruff
  Update pre-commit-hooks and add check-toml
  Change log for 38.2
  Fix the HiveSlimEventLoopMgr exit code when there is a failure in the algorithms
  Deprecate ApplicationMgr.StalledEventMonitoring property
  ...

+++ GaudiHive/src/CPUCruncher.cpp
@@ -59,7 +59,7 @@ StatusCode CPUCruncher::initialize() {
   }
 
   // if an algorithm was setup to sleep, for whatever period, it effectively models CPU-blocking behavior
-  if ( std::abs( m_sleepFraction ) > std::numeric_limits<float>::epsilon() ) setBlocking( true );
+  if ( std::abs( m_sleepFraction ) > std::numeric_limits<double>::epsilon() ) setBlocking( true );
 
   // This is a bit ugly. There is no way to declare a vector of DataObjectHandles, so
   // we need to wait until initialize when we've read in the input and output key
@@ -108,7 +108,7 @@ StatusCode CPUCruncher::execute() // the execution of the algorithm
 
   if ( m_loader && !m_declAugmented ) declareRuntimeRequestedOutputs();
 
-  float crunchtime;
+  double crunchtime;
 
   if ( m_local_rndm_gen ) {
     /* This will disappear with a thread safe random number generator service.
